CRUD de usuário finalizado

+++ src/main/java/br/bunny/controller/UserController.java
@@ -5,15 +5,17 @@ import br.bunny.model.User;
 import br.bunny.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.io.Serializable;
 import java.util.List;
 
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/users")
-public class UserController {
+public class UserController implements Serializable {
 
     private final UserService service;
 
@@ -35,8 +37,23 @@ public class UserController {
         }
     }
 
+    @PatchMapping  (path = "/{id}")
+    public ResponseEntity<UserDTO> activateOrDesactivateUser(@PathVariable("id") Long id, boolean ativo){
+        if(!service.existsById(id))
+            return ResponseEntity.notFound().build();
+        try {
+            service.activateOrDeactivate(id);
+            return ResponseEntity.noContent().build();
+        }catch (Exception e){
+            return ResponseEntity.internalServerError().build();
+        }
+    }
+
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteUser(@PathVariable("id") Long id){
+        if(!service.existsById(id))
+            return ResponseEntity.notFound().build();
+
         try{
             service.deleteById(id);
             return ResponseEntity.ok().build();
